slight restructuring for module passes and their tests
