+++ docs/source/modules/api/analysis/quantization.rst
@@ -1,5 +1,5 @@
 chop.passes.graph.calculate\_avg\_bits\_mg\_analysis\_pass
-===================================================
+==========================================================
