+++ docs/source/modules/api/passes.rst
@@ -1,16 +1,56 @@
-Chop Passes
+Passes
 ============================
+
 
 All passes, no matter analysis or transform, take a standard form:
 
 .. code-block:: python 
 
   # pass_args is a dict
-  def pass(mg, pass_args):
+  def pass(m, pass_args):
       ...
   # info a a dict
-  return mg, info
+  return m, info
 
+
+
+Mase Module Analysis Passes
+----------------------------
+
+.. list-table:: MASE module-level analysis passes 
+  :widths: 20 40 40
+  :header-rows: 1
+
+  * - Pass Name
+    - Usage Example
+    - Summary
+  * - :py:meth:`~chop.passes.module.analysis.quantize.calculate_avg_bits_module_analysis_pass`
+    - `test_calcualte_avg_bits_module <https://github.com/DeepWok/mase/blob/main/test/passes/module/analysis/quantize/test_calcualte_avg_bits_module.py>`_
+    - Analyzes the averaged bitwidth of a given module.
+
+.. toctree::
+  :maxdepth: 2
+
+  module_analysis/quantization
+
+Mase Module Transform Passes
+----------------------------
+
+.. list-table:: MASE module-level transform passes 
+  :widths: 20 40 40
+  :header-rows: 1
+
+  * - Pass Name
+    - Usage Example
+    - Summary
+  * - :py:meth:`~chop.passes.module.transforms.quantize.quantize_module_transform_pass`
+    - `test_module_quantize <https://github.com/DeepWok/mase/blob/main/test/passes/module/transforms/quantize/test_quantize_module.py>`_
+    - Apply quantization transformation to the given nn.Module
+
+.. toctree::
+  :maxdepth: 2
+
+  module_transform/quantization
 
 
 MaseGraph Analysis Passes
@@ -66,20 +106,32 @@ MaseGraph Analysis Passes
     - `test_hook_inspection_analysis_pass <https://github.com/DeepWok/mase/blob/main/test/passes/graph/anlysis/pruning/test_hook_inspect.py>`_
     - Provide hook information of the modules
   * - :py:meth:`~chop.passes.graph.analysis.runtime.runtime_analysis.runtime_analysis_pass`
-    - tbd
+    - fill me
     - Perform runtime analysis on the given graph (MaseGraph, TensorRT, ONNX models)
+  * - :py:meth:`~chop.passes.graph.analysis.verify.verify.verify_metadata_analysis_pass`
+    - fill me
+    - fill me
+  * - :py:meth:`~chop.passes.graph.analysis.verify.verify.verify_common_metadata_analysis_pass`
+    - fill me
+    - fill me
+  * - :py:meth:`~chop.passes.graph.analysis.verify.verify.verify_software_metadata_analysis_pass`
+    - fill me
+    - fill me
+  * - :py:meth:`~chop.passes.graph.analysis.verify.verify.verify_hardware_metadata_analysis_pass`
+    - fill me
+    - fill me
 
 .. toctree::
   :maxdepth: 2
 
-  analysis/add_metadata
   analysis/init_metadata
+  analysis/add_metadata
   analysis/report
   analysis/statistical_profiler
-  analysis/verify
   analysis/quantization
   analysis/pruning
   analysis/runtime
+  analysis/verify
 
 
 MaseGraph Transform Passes
@@ -138,10 +190,6 @@ MaseGraph Transform Passes
     - Apply TensorRT fine tune to the given graph for quantization aware training
     - fille me
 
-.. toctree::
-	:maxdepth: 1
-
-	transform/quantize_further_doc
 
 .. toctree::
   :maxdepth: 2
@@ -153,27 +201,35 @@ MaseGraph Transform Passes
   transform/tensorrt
 
 
-.. MaseGraph Interface Passes
-.. --------------------------
 
-.. .. list-table:: A summary of all MaseGraph interface passes 
-..   :widths: 25 75
-..   :header-rows: 1
+MaseGraph Interface Passes
+--------------------------
 
-..   * - Pass Name
-..     - Summary
-..   * - :py:meth:`~chop.passes.graph.interface.save_and_load.load_mase_graph_interface_pass`
-..     - fill me
-..   * - :py:meth:`~chop.passes.graph.interface.save_and_load.save_mase_graph_interface_pass`
-..     - fill me
-..   * - :py:meth:`~chop.passes.graph.interface.save_and_load.save_node_meta_param_interface_pass`
-..     - fill me
-..   * - :py:meth:`~chop.passes.graph.interface.save_and_load.load_node_meta_param_interface_pass`
-..     - fill me
-..   * - :py:meth:`~chop.passes.graph.interface.tensorrt.quantize.tensorrt_engine_interface_pass`
-..     - Converts the given graph to a TensorRT engine model
-..   * - :py:meth:`~chop.passes.graph.interface.onnxrt.onnx_runtime.onnx_runtime_interface_pass`
-..     - Converts the given graph to a ONNXRuntime model
+.. list-table:: A summary of all MaseGraph interface passes 
+  :widths: 20 40 40
+  :header-rows: 1
+
+  * - Pass Name
+    - Usage Example
+    - Summary
+  * - :py:meth:`~chop.passes.graph.interface.save_and_load.load_mase_graph_interface_pass`
+    - fill me
+    - fill me
+  * - :py:meth:`~chop.passes.graph.interface.save_and_load.save_mase_graph_interface_pass`
+    - fill me
+    - fill me
+  * - :py:meth:`~chop.passes.graph.interface.save_and_load.save_node_meta_param_interface_pass`
+    - fill me
+    - fill me
+  * - :py:meth:`~chop.passes.graph.interface.save_and_load.load_node_meta_param_interface_pass`
+    - fill me
+    - fill me
+  * - :py:meth:`~chop.passes.graph.interface.tensorrt.quantize.tensorrt_engine_interface_pass`
+    - Converts the given graph to a TensorRT engine model
+    - fill me
+  * - :py:meth:`~chop.passes.graph.interface.onnxrt.onnx_runtime.onnx_runtime_interface_pass`
+    - Converts the given graph to a ONNXRuntime model
+    - fill me
 
 .. toctree::
   :maxdepth: 2
@@ -182,5 +238,13 @@ MaseGraph Transform Passes
   interface/tensorrt
   interface/onnxrt
 
+
+Additional Documentation
+------------------------
+
+.. toctree::
+	:maxdepth: 1
+
+	transform/quantize_further_doc
 
 .. test-results:: ../../test/report.xml
