+++ src/chop/passes/module/analysis/__init__.py
@@ -1,1 +1,1 @@
-from .quantization import calculate_avg_bits_module_analysis_pass
+from .quantize import calculate_avg_bits_module_analysis_pass
