+++ src/chop/passes/module/transforms/quantize/__init__.py
@@ -1,2 +1,1 @@
-from .autosharding import resharding_transform_pass
 from .quantize import quantize_module_transform_pass
